Extract Todo route logic to handler file
